Trying and mostly failing to wrangle keyword args in Racket
A lot of code in the book passes keyword arguments around like
data. Common Lisp does this naturally, but not so much in Racket.
I would probably be better off using regular symbol/value pairs
in a list or associative lists or something. I might switch to
that next and get on with more interesting things.